Add Core SDK podspec generation through Fastlane
Fastlane is added along with an action that generates a podspec file
from a Podspec class. This class accepts a `Hash` with the podspec
information, and also provides some default values that are present
also on our Widgets SDK. Next step will be to integrate this into a
Bitrise workflow and have it triggered when the Core SDK is released.

MOB-1504