Add Core SDK podspec generation through Fastlane
Fastlane is added with a script taken from the Core SDK, that
generates a podspec from a previously generated template. The script
generates a new branch, generates the podspec with the version number,
creates a branch with the file, and then Fastlane uses that to create
a pull request in this repository.

MOB-1504