Create a new podspec on Widgets SDK release
Generate a new podspec for the Widgets SDK through Fastlane. A new
script has been added. which generates a new branch, generates the
podspec with the version number, creates a branch with the file, and
then Fastlane uses that to create a pull request in this repository.
Also, the creation of the Core SDK has been changed so that whenever
the Core SDK podspec is created, the Widgets SDK podspec template is
changed to reflect the new version too.

MOB-1511