Update `bitrise.yml` with web contents
Last time I did this, I forgot to then go to Bitrise and activate the
option to start having the `yml` file only on the repository, so now
I need to do it again. The latest change is that now there is a
workflow to add a podspec to the widgets too.

MOB-1517